Python Scripts/requirements.txt
google-generativeai
pillow
requests
pandas
